requirements.txt
pytorch-ignite
fire==0.4.0
lpips
pytorch_msssim==0.2.1
nilearn==0.9.0
tensorboard==2.8.0
pandas==1.1.5
nibabel==3.2.2
moviepy==1.0.3
performer-pytorch==1.1.4
deepspeed==0.4.2
mpi4py==3.1.3
x-transformers==0.25.4
monai==1.3.0
xformers==0.0.26.post1


